backend/requirements.txt
openai
python-dotenv
pandas
jupyter_client
ipykernel
